more refactor/moving of files
