+++ tests/test_core.py
@@ -9,7 +9,7 @@ import pytest
 
 # local
 import wcwidth
-from wcwidth._width import _WIDTH_FAST_PATH_MIN_LEN
+from wcwidth.width import _WIDTH_FAST_PATH_MIN_LEN
 
 _wcwidth_module = sys.modules['wcwidth.wcwidth']
 # local
